Add player state, bug fixes
- Refactoring and bug fixes to game memory reading

+++ src/utils/number_conversions.h
@@ -8,8 +8,8 @@ extern "C"
 
 #include <stdint.h>
 
-uint32_t big32_to_little(const char * const buf);
-uint32_t big16_to_little(const char * const buf);
+int32_t big32_to_little(const char * const buf);
+int16_t big16_to_little(const char * const buf);
 uint64_t ps3_address_to_x64(const uint32_t address);
 float big32_to_little_float(const char * const buf);
